requirements.txt
anyio==3.4.0
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Django==4.0.1
fastapi==0.71.0
fonttools==4.28.5
h11==0.12.0
idna==3.3
image==1.5.33
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.0
packaging==21.3
Pillow==9.0.0
pydantic==1.9.0
pyparsing==3.0.6
python-dateutil==2.8.2
qrcode==7.3.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.2
starlette==0.17.1
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
uvicorn==0.16.0
